backport quote when resetting a GUC and resolve empty values to 6x (#12972)
When a string guc such as search_path is set to '', Then some command such as create an extension with schema xxx, will add the changed guc to restore_list, before the next command begins to execute, call restore_guc_to_QE to restore changed guc. the restoring string is "set search_path to ;", then the string will be dispatched to QEs, QE elog an error, QD gets the error, and call ReThrowError to re-throw the error, errordata_stack_depth++, however, when we catch the error did not call FlushErrorState() to reset errordata_stack_depth.
when we execute other commands, after all, the errordata_stack_depth is not -1, The return value of elog_geterrcode shows there is an error.

related commits:320af40 and c4f03b8 and 620bdbd
related issue: #12900

tests: current extensions set search_path to public, so we can not reproduce this issue, but if I modify SQL for extension, It will be reproduced, reproduce step please see related issue:12900.